+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.5 (unreleased)
 
+- Added tensor classes - `FloatTensor`, `LongTensor`, etc
 - Improved modules
 
 ## 0.1.4 (2019-12-01)